Added AppNavLaunch to include Launch on XL breakpoint
